requirements.txt
# Core ML/DL Frameworks
torch>=2.1.0
torchvision>=0.16.0
tensorflow>=2.15.0
scikit-learn>=1.3.2
xgboost>=2.0.3

# Transformers & LLMs
transformers>=4.36.0
anthropic>=0.18.1
openai>=1.10.0
langchain>=0.1.4
langchain-community>=0.0.16
sentence-transformers>=2.3.1

# Vector Database & RAG
chromadb>=0.4.22
faiss-cpu>=1.7.4
pypdf>=4.0.1

# Time Series Forecasting
pytorch-forecasting>=1.0.0
darts>=0.27.0
prophet>=1.1.5
statsmodels>=0.14.1

# Geospatial & Satellite
earthengine-api>=0.1.384
rasterio>=1.3.9
geopandas>=0.14.2
shapely>=2.0.2
folium>=0.15.1
xarray>=2024.1.0
netCDF4>=1.6.5

# Graph Neural Networks
torch-geometric>=2.4.0
networkx>=3.2.1

# Computer Vision
opencv-python>=4.9.0
Pillow>=10.2.0

# Data Processing
pandas>=2.1.4
numpy>=1.26.3
scipy>=1.11.4

# API & Backend
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
pydantic-settings
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-generativeai==0.3.2

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
redis>=5.0.1
alembic>=1.13.1

# HTTP & Web Scraping
requests>=2.31.0
aiohttp>=3.9.1
beautifulsoup4>=4.12.3
selenium>=4.17.0

# Visualization
matplotlib>=3.8.2
seaborn>=0.13.1
plotly>=5.18.0

# UI Frameworks
streamlit>=1.30.0
gradio>=4.16.0

# WhatsApp Integration
twilio>=8.11.1

# Speech & Audio
openai-whisper
pydub>=0.25.1
elevenlabs>=0.2.27

# Monitoring & Logging
loguru>=0.7.2
python-json-logger>=2.0.7

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.1
joblib>=1.3.2
pytz>=2023.3
python-dateutil>=2.8.2

# Deployment
gunicorn>=21.2.0
docker>=7.0.0
kubernetes>=29.0.0

# Security
cryptography>=42.0.0
pyjwt>=2.8.0